api-gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
pyjwt==2.10.1
httpx==0.28.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
gunicorn==23.0.0
